Microservice-LLM/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
langchain
langchain-openai
python-dotenv
mysql-connector-python
celery[redis]